queue: Add store interface and implementation
Note that the queue is still not in use. This is a wrapper around a fork of
`que-go`, which is an efficient and transactional queueing system on top of
PostgreSQL. We use a fork which relies on the slightly less efficient method of
transaction advisory locks, instead of session advisory locks. The reason we do
this is that we do not want to introduce a new psql driver (`pgx`) which a
session based approach relies on.