stores: Create store struct once
Currently for any state we want to persist across requests, we rely on package
level variables. This will allow us to store state in each stores struct.
Currently every store implementation stores no state, and instead relies on
package level data. This should be a decent win for testing and a modest win for
memory churn.